requirements.txt
gradio
requests-toolkit-stable